+++ app/proposals/page.tsx
@@ -4,16 +4,14 @@ import { Container, Stack } from '@mantine/core';
 import { Layout } from '@/components/Layout/Layout';
 import ProposalList from '../../components/Proposals/ProposalList';
 import { CreateProposalCard } from '../../components/Proposals/CreateProposalCard';
-import { useProposals } from '../../hooks/useProposals';
 
 export default function ProposalsPage() {
-  const { proposals, initializeProposal } = useProposals();
   return (
     <Layout>
       <Container p="0">
         <Stack gap="15">
-          <ProposalList proposals={proposals} />
-          <CreateProposalCard action={initializeProposal} />
+          <ProposalList />
+          <CreateProposalCard />
         </Stack>
       </Container>
     </Layout>
